refactor: Adjusts Homepage vs Join Arena routes
